Updated how we represent state to match spec
